requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
transformers==4.37.2
torch==2.2.0
accelerate==0.27.2
python-multipart==0.0.9 